updated to remove mpi4py etc
