+++ scripts/example.py
@@ -1,18 +1,20 @@
+from time import time
+times={'t0' : time()}
+
 import jax
 import lpt4py as lpt
-from mpi4py import MPI
 import argparse
 import sys
-from time import time
+from jax.experimental.multihost_utils import sync_global_devices
 
 jax.config.update("jax_enable_x64", True)
 
 def myprint(*args,**kwargs):
     print("".join(map(str,args)),**kwargs);  sys.stdout.flush()
 
-def _profiletime(task_tag, step, times, comm=None, mpiproc=0):
-    if comm is not None:
-        comm.Barrier()
+def _profiletime(task_tag, step, times, parallel=False, host_id=0):
+    if parallel:
+        sync_global_devices('profiletime')
 
     dt = time() - times['t0']
     if step in times.keys():
@@ -21,7 +23,7 @@ def _profiletime(task_tag, step, times, comm=None, mpiproc=0):
         times[step] = dt
     times['t0'] = time()
 
-    if mpiproc!=0:
+    if host_id!=0:
         return times
 
     if task_tag is not None:
@@ -32,45 +34,37 @@ def _profiletime(task_tag, step, times, comm=None, mpiproc=0):
 
     return times
 
-times={'t0' : time()}
-
+dN = 128
+dseed = 13579
+dinfile = "__noise__"
 parser = argparse.ArgumentParser(description='Commandline interface to lpt4py example')
-parser.add_argument('--N',     type=int, help='grid dimention [default = 512]', default=512)
-parser.add_argument('--seed',  type=int, help='random seed [default = 13579]',  default=13579)
-parser.add_argument('--ityp',  type=str, help='lpt input type [default = delta]',  default='delta')
+parser.add_argument('--N',     type=int, help=f'grid dimention [default = {dN}]', default=dN)
+parser.add_argument('--seed',  type=int, help=f'noise with random seed when infile = "{dinfile}" [default = {dseed}]',  default=dseed)
+parser.add_argument('--infile',type=str, help=f'lpt input filename [default = "{dinfile}"]',  default=dinfile)
 
 args = parser.parse_args()
 
-N    = args.N
-seed = args.seed
-ityp = args.ityp
+N      = args.N
+seed   = args.seed
+infile = args.infile
 
-parallel = False
-nproc    = MPI.COMM_WORLD.Get_size()
-mpiproc  = MPI.COMM_WORLD.Get_rank()
-comm     = MPI.COMM_WORLD
-task_tag = "MPI process "+str(mpiproc)
-
-if MPI.COMM_WORLD.Get_size() > 1: parallel = True
-
-if not parallel:
-    cube = lpt.Cube(N=N,partype=None)  
+if jax.distributed.is_initialized():
+    parallel = True
+    host_id = jax.process_index()
 else:
-    jax.distributed.initialize()
-    cube = lpt.Cube(N=N)
-times = _profiletime(None, 'initialization', times, comm, mpiproc)
+    parallel = False
+    host_id = 0
 
-#### NOISE GENERATION
-delta = cube.generate_noise(seed=seed)
-times = _profiletime(None, 'noise generation', times, comm, mpiproc)
+if infile == dinfile:
+    box = lpt.Box(N=N,parallel=parallel,delta=None,seed=seed)
+else:
+    box = lpt.Box(N=N,parallel=parallel,delta=infile) 
 
-#### NOISE CONVOLUTION TO OBTAIN DELTA
-delta = cube.noise2delta(delta)
-times = _profiletime(None, 'noise convolution', times, comm, mpiproc)
+times = _profiletime(None, 'initialization', times, parallel, host_id)
 
-#### 2LPT DISPLACEMENTS FROM EXTERNAL (WEBSKY AT 768^3) DENSITY CONTRAST
-cube.slpt(infield=ityp,delta=delta)
-times = _profiletime(None, '2LPT', times, comm, mpiproc)
+# LPT displacements
+box.slpt()
+times = _profiletime(None, '2LPT', times, parallel, host_id)
 
 # LPT displacements are now in
 #   cube.s1x
